tools/req-parser/pyproject.toml
[project]
name = "req-parser"
version = "0.0.0"
description = "A requirement parser for McGill courses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "arrg>=0.1.0",
  "beautifulsoup4>=4.13.3",
  "openai>=1.66.3",
  "python-dotenv>=1.0.1",
]

[tool.ruff]
indent-width = 2
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20
indent-style = "space"
quote-style = "single"
